+++ sheriff/src/Sheriff/Rules.hs
@@ -27,10 +27,10 @@ showRule :: Rule
 showRule = FunctionRuleT $ FunctionRule "ShowRule" ["show"] 1 [] stringifierFns textTypesBlocked textTypesToCheck showRuleSuggestions showRuleExceptions [] ["*"] []
 
 infiniteRecursionRule :: Rule
-infiniteRecursionRule = InfiniteRecursionRuleT infiniteRecursionRuleT
+infiniteRecursionRule = InfiniteRecursionRuleT defaultInfiniteRecursionRuleT
 
-infiniteRecursionRuleT :: InfiniteRecursionRule 
-infiniteRecursionRuleT = defaultInfiniteRecursionRule {infinite_recursion_rule_name = "Infinite Recursion", infinite_recursion_rule_fixes = ["Remove the infinite recursion.", "Add a base case check.", "Pass the modified value to function arguments."]}
+defaultInfiniteRecursionRuleT :: InfiniteRecursionRule 
+defaultInfiniteRecursionRuleT = defaultInfiniteRecursionRule {infinite_recursion_rule_name = "Infinite Recursion", infinite_recursion_rule_fixes = ["Remove the infinite recursion.", "Add a base case check.", "Pass the modified value to function arguments."]}
 
 noKVDBRule :: Rule
 noKVDBRule = FunctionRuleT $ FunctionRule "ART KVDB Rule" ["runKVDB"] 0 [] [] [] [] ["You might want to use some other wrapper function from `EulerHS.Extra.Redis` module.", "For e.g. - rExists, rDel, rGet, rExpire, etc."] [] [] ["*"] []